Move the get-notification-orders.js to API folder
